LayoutTests/fast/exclusions/shape-outside-floats/shape-outside-floats-ellipse-margin-left-expected.txt

PASS elementRect('s1').top is 0
PASS elementRect('s1').left is 347
PASS elementRect('s2').top is 20
PASS elementRect('s2').left is 390
PASS elementRect('s3').top is 40
PASS elementRect('s3').left is 417
The left edges of the three black squares should follow the outer ellipse boundary and each square should appear on a subsequent line.

X
X
X
